fix potential integer overflow in depths calculation [build_me]

+++ src/depth.h
@@ -27,8 +27,8 @@
 #include "zlib.h"
 
 void *init_depths(void);
-void add_read_to_depths(const bam1_t *aln, const hts_pos_t qlen, void *depths, int32_t *hist, const int depths_max);
-void purge_and_reset_depths(void *depths, int32_t *hist, const int depths_max);
+void add_read_to_depths(const bam1_t *aln, const hts_pos_t qlen, void *depths, int64_t *hist, const int depths_max);
+void purge_and_reset_depths(void *depths, int64_t *hist, const int depths_max);
 void destroy_depths(void *depths);
 void *init_bedGraph(const params_t *params);
 void purge_and_reset_bedGraph(gzFile bgfile, void *bedGraph, const char *chr);
